Correcting parsing logic
As far as I can tell we always want this (and other functions) to have
the offset set on the first letter of the next datatype.  Also, we
don't want to increment more than one because we parse every field
every time.  This means that if we have an empty field "::" we don't
want to continue while it is still the separator character, then
the fields will be off.