Ignoring problems when pclose fails
unscd ignores SIGCHLD signals using code like this:

    /* Avoid creating zombies */
    signal(SIGCHLD, SIG_IGN);

That causes problems after we read the line of input and close the
process.  `pclose(fp)` returns -1 and `errno` is set to 10 (ECHILD).
Solutions include temporarily setting the interrupt handler back to a
default value or looking for this particular problem and ignoring it.
I've chosen the second because I don't want to possibly mess with the
parent program and let it get other SIGCHLD signals.