Must resort to signal handling
Turns out that the old code worked better in the success case but
utterly fell on its face when SIGCHLD was ignored and the exec script
returned a non-zero return code.

So, now we save the old signal handler, reset to default, run our stuff,
then restore things to the way they were.  This still only breaks on the
pclose call, but I wrap the entire three lines.